+++ crates/sui-bridge/Cargo.toml
@@ -48,6 +48,7 @@ shared-crypto.workspace = true
 backoff = { version = "0.4.0", features = [
   "futures",
 ] }
+enum_dispatch.workspace = true
 sui-json-rpc-api.workspace = true
 
 [dev-dependencies]
